flag for layers to indicate that image requests are fast to compute
The direct flag indicates that request to a layer's image source are
fast to compute, and that the viewer should wait for the computation
of these images in the GUI thread instead of off-loading the work to
a worker thread which increases the latency.

The manual test in tests/direct_layer_manual.py indicates the useful-
ness of this.