Implement a clickable layer.
A layer is now informed whether it is active or not (active means it is selected in the layerstack).
If a clickable layer is active, it switches the interpreter to a custom interpreter that intercepts right click
events and forwards them to a given functor.

The new ClickableColortableLayer uses this method.

In api.Viewer, the addClickableSegmentationLayer function adds a layer on which objects can be shown/hidden
via right click. Each object will get a unique, random color.