Workaround for performance bottleneck in ShortcutManager.
Whenever a new keyboard shortcut is registered, the ShorcutManager class
traverses the whole database to find and remove expired shortcuts.
This is especially expensive in the case where several input files are
open in ilastik and a new viewer is initialized for each file. Each of
these viewers register several shortcuts of their own.

Keeping a reverse map from the parent widgets to shortcuts belonging to
that widget speeds up the _purgeDeletedShortcuts() method to make this
use case with several input lanes in ilastik workable again.

This is a quick workaround. We should rethink the design of the
ShortcutManager class.