Do not remove missing layers in LayerDelegate
For some reason, `LayerDelegate.handleRemovedRows` could be called when
the model's data does not contain an item for the `EditRole`.

Also, `QAbstractItemModel.rowsAboutToBeRemoved` signal has an inclusive
right boundary, so parameters have been renamed to `first` and `last`,
which is consistent with the [Qt documentation](https://doc.qt.io/qt-5/qabstractitemmodel.html#rowsAboutToBeRemoved).

Closes https://github.com/ilastik/ilastik/issues/1997.