Move special case for mouse events to HUD (#258)
* Inherit all interpreters from InterpreterABC

Previously, interpreters relied on a custom `__subclasshook__` for
`isinstance` checks. Now, each interpreter is a subclass of
`InterpreterABC`, which is a proper abstract base class.
This change should make it easier to trace class relationships.

* Move special case for mouse events to HUD

Previously, EventSwtich installed itself as an event filter to both
ImageView2D (which is QGraphicsView) and it's viewport because a single
event filter should be able to catch all events (see https://stackoverflow.com/questions/2445997/qgraphicsview-and-eventfilter
for details). This made it hard to understand how events interact with
different event filters.

Another solution described in one of the anwsers in the link above is to
remove focus proxy from QGraphicsView. QAbstractScrollArea (superclass
of QGraphicsView) installs itself as a focus proxy to it's viewport
in order to be able to handle scrolling via keyboard. Since our
application doesn't need that, it is safe to remove the focus proxy and
simplify event handling.

However, HUD is currently displayed as a sibling of QGraphicsView
viewport, which causes problems when a user wants to do something with
a mouse in a HUD area that doesn't contain buttons, but has a usable
area underneath. We solve that problem by sending all mouse events from
HUD to the viewport of QGraphicsView.

Closes https://github.com/ilastik/ilastik/issues/2091.