fixes crash related to Overview3d
With a combination of Python>=3.7.4 and PyQt5<5.13.1 ilastik will crash when
instantiating the `Overview3d` object.

The problem are properties; In `PyQt5` during initialization of the Gui Widget
it calls `PyObject_GetAttr` on all class attributes. With some of them being
properties, they execute the code. But at this point (during init) the object
is not properly initialized. The error bubbles up since Python 3.7.4.

This commit removes properties in favour of getters/setters.