NOMERGE: one-off disable raising of tile fetch exceptions for release
Those highlight often temporary problems during short periods which are resolved
once slots are correctly set.

This is only for the 1.4.1 release, these temporary errors can be found in the
log files with detailed backtraces if needed.

f3ac824a3e8238b34f16dba615940665d922ebb3 is partially reverted to the state of
2b09109b478a7f30787d9cecbad01ce82812a671 .