Add layer priority
On layer creation a new integer `priority` argument can be used in ilastik to
prefer certain layers to be rendered/requested earlier.

E.g. Raw Data in general should be interesting to see first to
understand the current context better.

This will also enable the fixing of #272 (delayed display of labels
under load).